requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
moviepy>=1.0.3
pytesseract>=0.3.10
gtts>=2.3.0
tqdm>=4.65.0
